LPS-179214 Extract ScreenNavigationEntry methods from AccountEntryAddressesScreenNavigationCategory to AccountEntryAddressesScreenNavigationEntry, implement getCategoryKey method in it and return getCategoryKey in getEntryKey method
