+++ modules/apps/account/account-admin-web/src/main/java/com/liferay/account/admin/web/internal/frontend/taglib/servlet/taglib/AccountEntryAddressesScreenNavigationEntry.java
@@ -1,0 +1,82 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+package com.liferay.account.admin.web.internal.frontend.taglib.servlet.taglib;
+
+import com.liferay.account.admin.web.internal.constants.AccountScreenNavigationEntryConstants;
+import com.liferay.account.admin.web.internal.security.permission.resource.AccountEntryPermission;
+import com.liferay.account.constants.AccountActionKeys;
+import com.liferay.account.model.AccountEntry;
+import com.liferay.frontend.taglib.servlet.taglib.ScreenNavigationEntry;
+import com.liferay.portal.kernel.language.Language;
+import com.liferay.portal.kernel.model.User;
+import com.liferay.portal.kernel.security.permission.PermissionChecker;
+import com.liferay.portal.kernel.security.permission.PermissionCheckerFactoryUtil;
+
+import java.util.Locale;
+
+import org.osgi.service.component.annotations.Component;
+import org.osgi.service.component.annotations.Reference;
+
+/**
+ * @author Joao Victor Alves
+ */
+@Component(
+	property = "screen.navigation.entry.order:Integer=10",
+	service = ScreenNavigationEntry.class
+)
+public class AccountEntryAddressesScreenNavigationEntry
+	extends BaseAccountEntryScreenNavigationEntry {
+
+	@Override
+	public String getCategoryKey() {
+		return AccountScreenNavigationEntryConstants.CATEGORY_KEY_ADDRESSES;
+	}
+
+	@Override
+	public String getJspPath() {
+		return "/account_entries_admin/account_entry/addresses.jsp";
+	}
+
+	@Override
+	public String getLabel(Locale locale) {
+		return _language.get(locale, "addresses");
+	}
+
+	@Override
+	public boolean isVisible(User user, AccountEntry accountEntry) {
+		if (accountEntry.isNew()) {
+			return false;
+		}
+
+		PermissionChecker permissionChecker =
+			PermissionCheckerFactoryUtil.create(user);
+
+		if (AccountEntryPermission.contains(
+				permissionChecker, accountEntry.getAccountEntryId(),
+				AccountActionKeys.MANAGE_ADDRESSES) ||
+			AccountEntryPermission.contains(
+				permissionChecker, accountEntry.getAccountEntryId(),
+				AccountActionKeys.VIEW_ADDRESSES)) {
+
+			return true;
+		}
+
+		return false;
+	}
+
+	@Reference
+	private Language _language;
+
+}
